backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
websockets==13.1

google-genai==0.6.0